+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
   than relying on `Macro.to_string/1`. This change was advised by
   the Elixir team as `Macro.to_string/1` may not be entirely
   reliable.
+- Rendering of anonymous functions in typespecs.
 - Fix: Correctly insert spaces between function clauses when
   the function clause is preceeded by a comment.